+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   
 - Editor
   - Improve completion sorting, showing locals before functions and other completion items. #1158
+  - Fix hover to show current var definition docs instead of `def`/`defn`. #1157
 
 ## 2022.07.24-18.25.43